+++ popart_compiler/include/popart_compiler/CompilerOperationMacros.inc.hpp
@@ -20,7 +20,7 @@ OP_DECL(popart, dynamicupdate, dynamicupdate, AiGraphcoreOpset1.dynamicupdate, A
 OP_DECL(popart, dynamiczero, dynamiczero, AiGraphcoreOpset1.dynamiczero, ARG(INT_VEC,axes) ARG(INT_VEC,sizes) , BODY_ARG(axes) BODY_ARG(sizes) )
 OP_DECL(popart, dynamicadd, dynamicadd, AiGraphcoreOpset1.dynamicadd, ARG(INT_VEC,axes) ARG(INT_VEC,sizes) , BODY_ARG(axes) BODY_ARG(sizes) )
 OP_DECL(popart, sequenceslice, sequenceslice, AiGraphcoreOpset1.sequenceslice, ARG(INT,zeroUnused) , BODY_ARG(zeroUnused) )
-OP_DECL(popart, replicatedallreduce, replicatedallreduce, AiGraphcoreOpset1.replicatedallreduce, NONE, NONE)
+OP_DECL(popart, replicatedallreduce, replicatedallreduce, AiGraphcoreOpset1.replicatedallreduce, ARG(INT_VEC,commGroup) , BODY_ARG(commGroup) )
 OP_DECL(popart, l1loss, l1loss, AiGraphcoreOpset1.l1loss, ARG(FLOAT,lambda) ARG(INT,reduction) , BODY_ARG(lambda) BODY_ARG(static_cast<popart::ReductionType>(reduction)) )
 OP_DECL(popart, nllloss, nllloss, AiGraphcoreOpset1.nllloss, ARG(INT,reduction) ARG(INT,ignoreIndex) ARG(INT,inputIsLogProbability) , BODY_ARG(static_cast<popart::ReductionType>(reduction)) BODY_ARG(ignoreIndex) BODY_ARG(inputIsLogProbability) )
 OP_DECL(popart, identityloss, identityloss, AiGraphcoreOpset1.identityloss, ARG(INT,reduction) , BODY_ARG(static_cast<popart::ReductionType>(reduction)) )