+++ poptorch/include/poptorch/ImplicitCasting.hpp
@@ -19,10 +19,10 @@ namespace poptorch {
 enum class HalfFloatCasting { FloatDowncastToHalf = 0, HalfUpcastToFloat };
 
 void setHalfFloatCastingBehavior(HalfFloatCasting behavior);
-void setRunningVarianceAlwaysFloat(bool);
+void setRunningStatisticsAlwaysFloat(bool);
 HalfFloatCasting halfFloatCastingBehavior();
 
-bool runningVarianceAlwaysFloat();
+bool runningStatisticsAlwaysFloat();
 
 enum class ImplicitCast {
   None,
@@ -30,7 +30,7 @@ enum class ImplicitCast {
   ExceptFirst,
   ExceptSecond,
   ExceptThird,
-  ExceptFifth
+  ExceptFourthFifth
 };
 
 enum class ImplicitCastOutput { None, AsPromoted, AlwaysBool, AlwaysFloat };