+++ docs/user_guide/changelog.rst
@@ -22,6 +22,8 @@ New features
 - Add support for ``torch.scatter_add``
 - Add ``poptorch.Options.Precision.enableFloatingPointExceptions`` to control floating point exception behavior
 - Add option to log the number of IPU cycles used in executing the main graph
+- Deprecated ``runningVarianceAlwaysFloat`` in favour of ``runningStatisticsAlwaysFloat`` in ``poptorch.Options.Precision``,
+  as this new option computes both the running mean and variance in FP32 when this option is set to `True`.
 
 v2.0 (Poplar SDK 2.0)
 =====================